backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
pydantic
Pillow
pyjwt
bcrypt
motor
redis
pydantic[email]
